+++ test/unit/fuzz-tools.test.js
@@ -367,6 +367,7 @@ import * as dimeglioClubfootV387 from '../../lib/dimeglio-clubfoot-v387.js';
 import * as brodskyTonsilV388 from '../../lib/brodsky-tonsil-v388.js';
 import * as koosSchwannomaV389 from '../../lib/koos-schwannoma-v389.js';
 import * as knospAdenomaV390 from '../../lib/knosp-adenoma-v390.js';
+import * as hardyAdenomaV391 from '../../lib/hardy-adenoma-v391.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -679,6 +680,7 @@ const MODULES = {
   'brodsky-tonsil-v388.js': brodskyTonsilV388,
   'koos-schwannoma-v389.js': koosSchwannomaV389,
   'knosp-adenoma-v390.js': knospAdenomaV390,
+  'hardy-adenoma-v391.js': hardyAdenomaV391,
   'tb-testing.js': tbTesting,
 };